🔄 refactor(ui): extract responsive layout logic into useDashboardResponsiveLayouts
- Move layout hydration, persistence, breakpoint management, and memo into
  a dedicated composable for separation of concerns
- Cache widget breakpoint bounds in DashboardView computed instead of
  calling getWidgetBoundsForBreakpoint per GridItem prop binding
- Add delegation test verifying useDashboardWidgetOrder imports the new module
