+++ ui/src/views/dashboard/useDashboardWidgetOrder.ts
@@ -1,29 +1,16 @@
-import { computed, onScopeDispose, ref, watch } from 'vue';
-import {
-  DASHBOARD_LAYOUT_BREAKPOINTS,
-  type DashboardLayoutBreakpoint,
-  type PersistedLayoutItem,
-  type PersistedResponsiveLayoutMap,
-} from '../../preferences/schema';
+import { onScopeDispose, ref, watch } from 'vue';
 import { preferences } from '../../preferences/store';
 import { DASHBOARD_WIDGET_IDS, type DashboardWidgetId } from './dashboardTypes';
 import {
-  applyConstraints,
-  createDefaultLayoutForBreakpoint,
-  type WidgetLayoutItem,
-} from './dashboardWidgetLayout';
+  _rebuildLayoutsForOrderForTests,
+  createResponsiveLayoutsMemo,
+  useDashboardResponsiveLayouts,
+} from './useDashboardResponsiveLayouts';
 
-type ResponsiveWidgetLayouts = Partial<Record<DashboardLayoutBreakpoint, WidgetLayoutItem[]>>;
-
-const RESPONSIVE_BREAKPOINTS = DASHBOARD_LAYOUT_BREAKPOINTS as readonly DashboardLayoutBreakpoint[];
-const DEFAULT_BREAKPOINT: DashboardLayoutBreakpoint = 'lg';
+export { _rebuildLayoutsForOrderForTests, createResponsiveLayoutsMemo };
 
 function isDashboardWidgetId(value: unknown): value is DashboardWidgetId {
   return typeof value === 'string' && (DASHBOARD_WIDGET_IDS as readonly string[]).includes(value);
-}
-
-function isDashboardLayoutBreakpoint(value: unknown): value is DashboardLayoutBreakpoint {
-  return typeof value === 'string' && (RESPONSIVE_BREAKPOINTS as readonly string[]).includes(value);
 }
 
 function arraysEqual<T>(left: readonly T[], right: readonly T[]): boolean {
@@ -64,219 +51,6 @@ function sanitizeWidgetOrder(rawOrder: unknown): DashboardWidgetId[] {
   return sanitized;
 }
 
-function cloneLayoutItem(item: WidgetLayoutItem): WidgetLayoutItem {
-  return {
-    i: item.i,
-    x: item.x,
-    y: item.y,
-    w: item.w,
-    h: item.h,
-    minW: item.minW,
-    minH: item.minH,
-    maxW: item.maxW,
-    maxH: item.maxH,
-  };
-}
-
-function cloneLayout(layout: readonly WidgetLayoutItem[]): WidgetLayoutItem[] {
-  return layout.map(cloneLayoutItem);
-}
-
-/* v8 ignore next 12 -- short-circuit chain; every property is tested individually */
-function layoutItemsEqual(left: WidgetLayoutItem, right: WidgetLayoutItem): boolean {
-  return (
-    left.i === right.i &&
-    left.x === right.x &&
-    left.y === right.y &&
-    left.w === right.w &&
-    left.h === right.h &&
-    left.minW === right.minW &&
-    left.minH === right.minH &&
-    left.maxW === right.maxW &&
-    left.maxH === right.maxH
-  );
-}
-
-function layoutsShallowEqual(
-  left: readonly WidgetLayoutItem[] | undefined,
-  right: readonly WidgetLayoutItem[] | undefined,
-): boolean {
-  if (!left || !right || left.length !== right.length) {
-    return false;
-  }
-  return left.every((item, index) => layoutItemsEqual(item, right[index]!));
-}
-
-/** @internal Exported for testing only. */
-export function createResponsiveLayoutsMemo() {
-  let previousResult: ResponsiveWidgetLayouts = {};
-
-  return (layouts: ResponsiveWidgetLayouts): ResponsiveWidgetLayouts => {
-    let changed = false;
-    const nextResult: ResponsiveWidgetLayouts = {};
-
-    for (const breakpoint of RESPONSIVE_BREAKPOINTS) {
-      const source = layouts[breakpoint];
-      const previous = previousResult[breakpoint];
-
-      if (!source?.length) {
-        if (previous?.length) {
-          changed = true;
-        }
-        continue;
-      }
-
-      if (layoutsShallowEqual(source, previous)) {
-        nextResult[breakpoint] = previous;
-        continue;
-      }
-
-      nextResult[breakpoint] = cloneLayout(source);
-      changed = true;
-    }
-
-    if (!changed) {
-      return previousResult;
-    }
-
-    previousResult = nextResult;
-    return previousResult;
-  };
-}
-
-function stripLayout(layout: readonly WidgetLayoutItem[]): PersistedLayoutItem[] {
-  return layout.map((item) => ({
-    i: item.i,
-    x: item.x,
-    y: item.y,
-    w: item.w,
-    h: item.h,
-  }));
-}
-
-function isValidLayoutItem(value: unknown): value is WidgetLayoutItem {
-  if (!value || typeof value !== 'object') return false;
-  const item = value as Record<string, unknown>;
-  return (
-    isDashboardWidgetId(item.i) &&
-    typeof item.x === 'number' &&
-    typeof item.y === 'number' &&
-    typeof item.w === 'number' &&
-    typeof item.h === 'number'
-  );
-}
-
-function isLegacySingleColumnLayout(rawLayout: unknown): boolean {
-  return (
-    Array.isArray(rawLayout) &&
-    rawLayout.length > 0 &&
-    rawLayout.every((item) => isValidLayoutItem(item) && item.x === 0 && item.w === 1)
-  );
-}
-
-function createLayoutFromOrder(
-  order: readonly DashboardWidgetId[],
-  breakpoint: DashboardLayoutBreakpoint = DEFAULT_BREAKPOINT,
-): WidgetLayoutItem[] {
-  return createDefaultLayoutForBreakpoint(order, breakpoint);
-}
-
-function hydrateLayout(
-  order: readonly DashboardWidgetId[],
-  breakpoint: DashboardLayoutBreakpoint,
-  rawLayout: unknown,
-): WidgetLayoutItem[] {
-  const baseLayout = createLayoutFromOrder(order, breakpoint);
-  if (!Array.isArray(rawLayout)) {
-    return baseLayout;
-  }
-
-  const persisted = new Map<DashboardWidgetId, WidgetLayoutItem>();
-  for (const item of rawLayout) {
-    if (isValidLayoutItem(item)) {
-      persisted.set(item.i, { i: item.i, x: item.x, y: item.y, w: item.w, h: item.h });
-    }
-  }
-
-  return applyConstraints(
-    baseLayout.map((item) => {
-      const saved = persisted.get(item.i);
-      return saved ? { ...item, ...saved } : item;
-    }),
-    breakpoint,
-  );
-}
-
-function loadPersistedLayouts(order: readonly DashboardWidgetId[]): ResponsiveWidgetLayouts {
-  const layouts: ResponsiveWidgetLayouts = {};
-  const rawResponsiveLayouts = preferences.dashboard.gridLayouts;
-
-  /* v8 ignore next -- defensive: gridLayouts is always initialized as object */
-  if (rawResponsiveLayouts && typeof rawResponsiveLayouts === 'object') {
-    for (const breakpoint of RESPONSIVE_BREAKPOINTS) {
-      const candidate = (rawResponsiveLayouts as Record<string, unknown>)[breakpoint];
-      if (Array.isArray(candidate)) {
-        layouts[breakpoint] = hydrateLayout(order, breakpoint, candidate);
-      }
-    }
-  }
-
-  if (
-    Object.keys(layouts).length === 0 &&
-    Array.isArray(preferences.dashboard.gridLayout) &&
-    preferences.dashboard.gridLayout.length > 0
-  ) {
-    /* v8 ignore next -- legacy migration: tested via migration specs */
-    const legacyBreakpoint = isLegacySingleColumnLayout(preferences.dashboard.gridLayout)
-      ? 'sm'
-      : 'lg';
-    layouts[legacyBreakpoint] = hydrateLayout(
-      order,
-      legacyBreakpoint,
-      preferences.dashboard.gridLayout,
-    );
-  }
-
-  if (!layouts.lg) {
-    layouts.lg = createLayoutFromOrder(order, 'lg');
-  }
-
-  return layouts;
-}
-
-function serializeResponsiveLayouts(
-  layouts: ResponsiveWidgetLayouts,
-): PersistedResponsiveLayoutMap {
-  const result: PersistedResponsiveLayoutMap = {};
-  for (const breakpoint of RESPONSIVE_BREAKPOINTS) {
-    if (layouts[breakpoint]?.length) {
-      result[breakpoint] = stripLayout(layouts[breakpoint]!);
-    }
-  }
-  return result;
-}
-
-function rebuildLayoutsForOrder(
-  order: readonly DashboardWidgetId[],
-  layouts: ResponsiveWidgetLayouts,
-): ResponsiveWidgetLayouts {
-  const nextLayouts: ResponsiveWidgetLayouts = {};
-
-  for (const breakpoint of RESPONSIVE_BREAKPOINTS) {
-    if (layouts[breakpoint]?.length) {
-      nextLayouts[breakpoint] = hydrateLayout(order, breakpoint, layouts[breakpoint]);
-    }
-  }
-
-  if (!nextLayouts.lg) {
-    nextLayouts.lg = createLayoutFromOrder(order, 'lg');
-  }
-
-  return nextLayouts;
-}
-
-export const _rebuildLayoutsForOrderForTests = rebuildLayoutsForOrder;
-
 function getDragSource(event: DragEvent): DashboardWidgetId | null {
   const rawSource = event.dataTransfer?.getData('text/plain');
   return isDashboardWidgetId(rawSource) ? rawSource : null;
@@ -302,20 +76,19 @@ export function moveWidget(
 }
 
 export function useDashboardWidgetOrder() {
-  const getResponsiveLayouts = createResponsiveLayoutsMemo();
   const widgetOrder = ref<DashboardWidgetId[]>(
     sanitizeWidgetOrder(preferences.dashboard.widgetOrder),
   );
-  const currentBreakpoint = ref<DashboardLayoutBreakpoint>(DEFAULT_BREAKPOINT);
-  const layoutsByBreakpoint = ref<ResponsiveWidgetLayouts>(loadPersistedLayouts(widgetOrder.value));
-  const layout = ref<WidgetLayoutItem[]>(
-    cloneLayout(
-      /* v8 ignore next -- defensive fallback for missing breakpoint */
-      layoutsByBreakpoint.value[currentBreakpoint.value] ??
-        createLayoutFromOrder(widgetOrder.value, currentBreakpoint.value),
-    ),
-  );
-  const responsiveLayouts = computed(() => getResponsiveLayouts(layoutsByBreakpoint.value));
+  const {
+    currentBreakpoint,
+    layout,
+    onBreakpointChanged,
+    persistDashboardLayouts,
+    rebuildLayoutsForCurrentOrder,
+    resetLayoutsToDefaults,
+    responsiveLayouts,
+    syncCurrentLayoutFromResponsiveLayouts,
+  } = useDashboardResponsiveLayouts({ widgetOrder });
   const gridInstanceKey = ref(0);
   const hiddenWidgets = ref<DashboardWidgetId[]>(
     sanitizeHiddenWidgets(preferences.dashboard.hiddenWidgets),
@@ -329,43 +102,14 @@ export function useDashboardWidgetOrder() {
     gridInstanceKey.value += 1;
   }
 
-  function syncCurrentLayoutIntoResponsiveLayouts(nextLayout: readonly WidgetLayoutItem[]) {
-    const normalized = hydrateLayout(widgetOrder.value, currentBreakpoint.value, nextLayout);
-    layoutsByBreakpoint.value = {
-      ...layoutsByBreakpoint.value,
-      [currentBreakpoint.value]: cloneLayout(normalized),
-    };
-    return normalized;
-  }
-
-  function persistDashboardLayouts(nextLayout: readonly WidgetLayoutItem[] = layout.value) {
-    const normalized = syncCurrentLayoutIntoResponsiveLayouts(nextLayout);
-    preferences.dashboard.widgetOrder = [...widgetOrder.value];
-    preferences.dashboard.gridLayouts = serializeResponsiveLayouts(layoutsByBreakpoint.value);
-    /* v8 ignore next -- defensive fallback: lg always exists */
-    preferences.dashboard.gridLayout = [...(preferences.dashboard.gridLayouts.lg ?? [])];
-    return normalized;
-  }
-
   function persistHiddenWidgets() {
     preferences.dashboard.hiddenWidgets = [...hiddenWidgets.value];
-  }
-
-  function syncCurrentLayoutFromResponsiveLayouts() {
-    layout.value = cloneLayout(
-      /* v8 ignore next -- defensive fallback for missing breakpoint */
-      layoutsByBreakpoint.value[currentBreakpoint.value] ??
-        createLayoutFromOrder(widgetOrder.value, currentBreakpoint.value),
-    );
   }
 
   function applyWidgetOrder(nextOrder: readonly DashboardWidgetId[]) {
     syncing = true;
     widgetOrder.value = [...nextOrder];
-    layoutsByBreakpoint.value = rebuildLayoutsForOrder(
-      widgetOrder.value,
-      layoutsByBreakpoint.value,
-    );
+    rebuildLayoutsForCurrentOrder();
     syncCurrentLayoutFromResponsiveLayouts();
     persistDashboardLayouts(layout.value);
     refreshGridInstance();
@@ -381,7 +125,7 @@ export function useDashboardWidgetOrder() {
         return;
       }
       syncing = true;
-      layoutsByBreakpoint.value = rebuildLayoutsForOrder(nextOrder, layoutsByBreakpoint.value);
+      rebuildLayoutsForCurrentOrder();
       syncCurrentLayoutFromResponsiveLayouts();
       persistDashboardLayouts(layout.value);
       refreshGridInstance();
@@ -450,27 +194,6 @@ export function useDashboardWidgetOrder() {
     return { order: widgetOrderIndex(widgetId) };
   }
 
-  function onBreakpointChanged(
-    breakpoint: DashboardLayoutBreakpoint,
-    nextLayout?: readonly WidgetLayoutItem[],
-  ) {
-    if (!isDashboardLayoutBreakpoint(breakpoint)) {
-      return;
-    }
-
-    currentBreakpoint.value = breakpoint;
-    const normalized = hydrateLayout(
-      widgetOrder.value,
-      breakpoint,
-      /* v8 ignore next -- defensive fallback when called without explicit layout */
-      nextLayout ?? layoutsByBreakpoint.value[breakpoint],
-    );
-    layoutsByBreakpoint.value = {
-      ...layoutsByBreakpoint.value,
-      [breakpoint]: cloneLayout(normalized),
-    };
-  }
-
   function onWidgetDragStart(widgetId: DashboardWidgetId, event: DragEvent) {
     draggedWidgetId.value = widgetId;
     if (event.dataTransfer) {
@@ -522,10 +245,7 @@ export function useDashboardWidgetOrder() {
       syncing = true;
       hiddenWidgets.value = hiddenWidgets.value.filter((id) => id !== widgetId);
       widgetOrder.value = sanitizeWidgetOrder([...widgetOrder.value, widgetId]);
-      layoutsByBreakpoint.value = rebuildLayoutsForOrder(
-        widgetOrder.value,
-        layoutsByBreakpoint.value,
-      );
+      rebuildLayoutsForCurrentOrder();
       syncCurrentLayoutFromResponsiveLayouts();
       persistDashboardLayouts(layout.value);
       refreshGridInstance();
@@ -546,17 +266,7 @@ export function useDashboardWidgetOrder() {
     syncing = true;
     hiddenWidgets.value = [];
     widgetOrder.value = [...DASHBOARD_WIDGET_IDS];
-    const nextLayouts: ResponsiveWidgetLayouts = {
-      lg: createLayoutFromOrder(widgetOrder.value, 'lg'),
-    };
-    if (currentBreakpoint.value !== 'lg') {
-      nextLayouts[currentBreakpoint.value] = createLayoutFromOrder(
-        widgetOrder.value,
-        currentBreakpoint.value,
-      );
-    }
-    layoutsByBreakpoint.value = nextLayouts;
-    syncCurrentLayoutFromResponsiveLayouts();
+    resetLayoutsToDefaults();
     persistHiddenWidgets();
     persistDashboardLayouts(layout.value);
     refreshGridInstance();
